ParticleDensityOpacityModule migrated from aowt.plugin
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/mmstd_datatools/trunk@43 f95d2973-9550-45a5-b795-f9f66c6b6f94

+++ ParticleListSelector.cpp
@@ -45,7 +45,7 @@ bool datatools::ParticleListSelector::manipulateData(
     inData.SetUnlocker(nullptr, false); // keep original data locked
                                         // original data will be unlocked through outData
 
-    if ((idx < 0) || (idx >= inData.GetParticleListCount())) {
+    if ((idx < 0) || (idx >= static_cast<int>(inData.GetParticleListCount()))) {
         outData.SetParticleListCount(0);
     } else {
         outData.SetParticleListCount(1);
